create mask from cortical ribbon, create stim.nii.gz

+++ replace_vtks.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
 for i in pial white inflated sphere
-do
-for hemi in lh rh
-do
-rm prf/surfaces/${hemi}.${i}.vtk
-cp /N/u/davhunt/Carbonate/Downloads/Bradley_prf/surfaces/${hemi}.${i}.gii prf/surfaces/${hemi}.${i}.gii
-mris_convert prf/surfaces/${hemi}.${i}.gii prf/surfaces/${hemi}.${i}.vtk
-done
+  do
+  for hemi in lh rh
+    do
+    if [ -f prf/surfaces/${hemi}.${i}.vtk ]; then rm prf/surfaces/${hemi}.${i}.vtk; fi
+    cp /N/u/davhunt/Carbonate/Downloads/Bradley_prf/surfaces/${hemi}.${i}.gii prf/surfaces/${hemi}.${i}.gii
+    mris_convert prf/surfaces/${hemi}.${i}.gii prf/surfaces/${hemi}.${i}.vtk
+  done
 done
